requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2025.11.12
channels==4.1.0
charset-normalizer==3.2.0
daphne==4.1.2
decorator==5.1.1
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
ebird-api==3.0.6
executing==1.2.0
idna==3.4
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
ptyprocess==0.7.0
pure-eval==0.2.2
psycopg2-binary==2.9.9
Pygments==2.15.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
shortuuid==1.0.13
oauthlib==3.2.2
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
Pillow==10.4.0
html2text==2024.2.26
beautifulsoup4==4.12.2
openai==1.3.0
markdown==3.5.1